requirements.txt
flask
pyjwt[crypto]
python-dotenv